Add additional check to see if user is authenticated in comment delete model

+++ news/views.py
@@ -90,7 +90,7 @@ def comment_delete(request, slug, comment_id):
     article = get_object_or_404(queryset, slug=slug)
     comment = get_object_or_404(Comment, pk=comment_id)
 
-    if comment.author == request.user:
+    if comment.author == request.user and request.user.is_authenticated:
         comment.delete()
         messages.add_message(request, messages.SUCCESS, 'Comment deleted!')
     else:
